// apps/web/lib/api/generated/models/listChatsParams.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * llame API
 * llame auth and domain API
 * OpenAPI spec version: 0.1
 */
import type { ListChatsArchived } from "./listChatsArchived";
import type { ListChatsPinned } from "./listChatsPinned";

export type ListChatsParams = {
  projectId?: string;
  archived?: ListChatsArchived;
  pinned?: ListChatsPinned;
};